Add different humanity change cases
Reading _aialignment from the SpawnGroup function to determine whether to add or subtract humanity from players for killing the unit.  This is done in a case structure to allow for more flexibility in adding unit alignments.  This would also allow for variable humanity gains for different unit types within the positive/negative alignments (think hostages).